+++ lib/super/controls/required.rb
@@ -8,16 +8,6 @@ module Super
       def model
         @actual.model
       end
-
-      # Configures which parameters could be written to the database. This is a
-      # required method
-      #
-      # @param params [ActionController::Parameters]
-      # @param action [ActionInquirer]
-      # @return [ActionController::Parameters]
-      def permitted_params(params, action:)
-        @actual.permitted_params(params, action: action)
-      end
     end
   end
 end