+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Additional "schema type" helpers on Display and Form
 * Schema guesser for `#display_schema`, `#form_schema`, and `#filter_schema`.
+* Guesser for Strong Parameters
 
 
 ## [0.0.7] - 2020-12-15